Put in linear data
karlosz pointed out that I was barking pretty far off the wrong
tree in terms of how to make analysis passes easier. SSA is nice,
but if things have only one _use_ we can easily tag them with a
type or whatnot, which SSA doesn't help with (because the same
variable could have different types in different branches, etc.)

Thus, the linear-datum.

This does break the local variable eliminator though. It can't be
done in the same way, as multiple readvars could be replaced with
the same datum, making that datum nonlinear.